add useProposalsList custom hook
