+++ app/components/views/GovernancePage/Proposals/ProposalsTab.jsx
@@ -1,4 +1,4 @@
-import { ProposalList } from "./ProposalsList/ProposalsList";
+import ProposalList from "./ProposalsList/ProposalsList";
 import PoliteiaDisabled from "./PoliteiaDisabled";
 import { useSelector, useDispatch } from "react-redux";
 import { setLastPoliteiaAccessTime } from "actions/WalletLoaderActions";
